requirements.txt
python-dotenv==1.0.0
requests==2.28.2
pytest==3.2.2
pytest-cov==2.5.1
pytest-dependency==0.3.2
flask